requirements.txt
selenium==4.25.0
pytest==8.3.3
pytest-dependency==0.6.0
python-dotenv==1.0.0